backended/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.1
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.330
langsmith==0.0.57
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.2
yt-dlp==2023.10.13
